Add ':addclause' command at the REPL
:addclause l n
will find the name 'n' on line 'l', which should be a type declaration,
and print a corresponding clause for it. :addclause! does the same, but
adds it to the first blank line after that in the file

vim script updated so that 'cd' does the corresponding thing (moving back
to find the type declaration before the cursor)